Reorganize project structure to match interop generator
- Create Helpers/ folder and move TypeMapper there (namespace: Helpers)
- Extract MappedType struct to Models/MappedType.cs (was inlined in TypeMapper)
- Fix TypeDeclaration namespace to match its Models/ folder location
- Update all using directives across the project

This aligns the project layout with the interop generator's convention of
separating utility/mapping classes (Helpers/) from data models (Models/).
